pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "atlas-mapping"
version = "0.1.0"
description = "MITRE ATLAS mapping utilities for YOLO-based maritime perception"
authors = [{name = "milliAmpere2"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = []

[project.scripts]
atlas-mitigation = "atlas_mapping.image_analyzer:main"

[project.optional-dependencies]
yolo = [
    "torch>=2.0.0",
    "pandas>=1.3",
]
art = [
    "adversarial-robustness-toolbox>=1.18.0",
    "numpy>=1.23",
    "pillow>=9.4",
]

[tool.setuptools]
packages = ["atlas_mapping"]
